Moved unrolled methods to a partial class, bench: 5949647
